+++ NEWS.md
@@ -3,6 +3,7 @@
 * Bug fix for HSVD function not working with frequency domain data.
 * Bug fix for TWIX affine calculation when not using power of two matrix
 dimensions.
+* Added ret_metab_only option to comb_coils.
 
 # spant 2.1.0
 * y-axis label can now be changed with a new argument (yaxis_lab) to